requirements.txt
aiohttp==3.6.2
aniso8601==7.0.0
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptg==0.2.post0
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.3
django-filter==2.2.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
get==2019.4.13
graphene==2.1.8
graphene-django==2.8.2
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.2
ipdb==0.12.2
ipykernel==5.2.0
ipython==7.10.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.1
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
post==2019.4.13
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
ptyprocess==0.6.0
public==2019.4.13
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.5.2
pyrsistent==0.16.0
python-dateutil==2.8.1
python-telegram-bot==12.4.2
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
rsa==4.0
Rx==1.6.1
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
Telethon==1.11.3
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==3.1.0
